drop darwin from official manifest, build local platform for devs
The plugin.json manifest has been "lying", in that it lists support for MacOS (`darwin`) even though we don't support or distribute that. It's necessary to allow developers to work locally.

Fix this by detecting if `MM_SERVICESETTINGS_ENABLEDEVELOPER` is set as truthy in the environment, and dynamically rewriting the manifest to clear the `Executables` and instead the platform-agnostic `Executable`. The server won't know what platform is supported, but will assume it's the right one and "just run it", working fine in development while leaving CI and production builds unchanged.

Then unlist darwin from the canonical manifest.